Stepan Roh fixed some problems with 64-bit machines

+++ src/midi.c
@@ -95,7 +95,7 @@ long midi_loop_end = -1;                        /* loop at this position */
 static int midi_semaphore = 0;                  /* reentrancy flag */
 static int midi_loaded_patches = FALSE;         /* loaded entire patch set? */
 
-static int midi_timer_speed;                    /* midi_player's timer speed */
+static long midi_timer_speed;                    /* midi_player's timer speed */
 static int midi_pos_speed;                      /* MIDI delta -> midi_pos */
 static int midi_speed;                          /* MIDI delta -> timer */
 static int midi_new_speed;                      /* for tempo change events */
@@ -412,7 +412,7 @@ int _midi_allocate_voice(int min, int max)
    int c;
    int layer;
    int voice = -1;
-   int best_time = LONG_MAX;
+   long best_time = LONG_MAX;
 
    if (min < 0)
       min = 0;
